Correct use of __slots__.
Turns out, slots should not be redefined in subclasses. Thus, MetaNode
adds an empty slot list when none is givem, and adds a _allslots
attribute containing all the inherited slot names.

Note that inheriting from two classes that define slots does not
appear to be possible. Fortunately, this is not presently needed.