New semantics for nsz
The previous semantics for nsz did not properly constrain the undef value produced
when op nsz returns +/-0. The new implementation uses a quantified boolean,
and selects +0 or -0 depending on its value.

Problem: this seems to cause Z3 to give up on some (valid) optimizations.

The old behavior is still available as the BrokenNSZ translator.

Also adds the OldNSZ translator, which excludes -0 using a well-defined condition.