Refactor alive.py to separate logic and status reporting
Refinement errors are now signaled by exceptions. All reporting state in
alive.py is handled by new class StatusReporter, which tracks what
information to report and to which output stream (stdout/stderr).

The new code to test a suite of opts is short enough that it has been
folded back into main(). This will change if other applications are
introduced.