Refactor alive.smtinterp.
Generalized OpHandler to evaluate args and pass them to the given
function. (It can also be used as a decorator.) Since the generalized
form uses Node.args() instead of examining fields directly, it
can be used widely. E.g., ZExtInst and ZExtCnxp use the same
OpHandler object.

Presumably, there is a performance hit for this, but it seems
minor.