subterms() generates values in post-order
Turns out, the post-order is more useful for reporting counter-examples,
and we couldn't get it by reversing the old pre-order. The new
subterms ensures that all terms are emitted after their subterms (if any).
The algorithm is asymptotically less efficient, because it creates
O(n) nested generators, but n is generally small so this shouldn't be
an issue.