Permit constant definitions in target
The target can now define new constants, referring only to values
defined in the source and constants defined in the target. These
constants are also in scope in the precondition.

Aside from making some opts shorter, this is important for
polymorphic constant expressions, as it provides a way to ensure
that all instances of an expression have the same type.