Represent feature results for test cases using three values
For a given test case, a feature may return ACCEPT, REJECT, or
UNSAFE. Both the feature and its negation are considered false
when it is UNSAFE, i.e., it is assumed that UNSAFE results are
avoided.

This means the boolean learner is no longer guaranteed to find
a CNF expression separating the good and bad vectors.