Restructure infer_preconditions, add precondition search
Instead of halting infer_preconditions_by_examples when a precondition
is unsound or incomplete, we mutate the example lists and throw
MoreExamples, signalling to i_p_b_e to reconstruct its feature
vectors.

This dovetails nicely with preconditions_satisfying_examples, which
never needs to consider a precondition twice.

Reconstructing the feature vectors is not ideal, but adding examples
to an already existing feature matrix seemed tricky, especially when
new positive examples may be unsafe for already-accepted features.